fix(bsp): remove -Wconversion warning

+++ bsp/boards/bsp_delay.c
@@ -12,22 +12,20 @@
 
 #include "bsp_delay.h"
 
-#include "main.h"
-
-static uint8_t fac_us = 0U;
-static uint32_t fac_ms = 0U;
+static uint8_t fac_us = 0;
+static uint32_t fac_ms = 0;
 
 void delay_init(void)
 {
-    fac_us = SystemCoreClock / 1000000U;
-    fac_ms = SystemCoreClock / 1000U;
+    fac_us = (uint8_t)(SystemCoreClock / 1000000);
+    fac_ms = SystemCoreClock / 1000;
 }
 
 void delay_us(uint16_t nus)
 {
-    uint32_t tcnt = 0U;
-    uint32_t tnow = 0U;
-    uint32_t ticks = nus * fac_us;
+    uint32_t tcnt = 0;
+    uint32_t tnow = 0;
+    uint32_t ticks = (uint32_t)nus * fac_us;
     uint32_t reload = SysTick->LOAD; /* SysTick Reload Value Register */
     uint32_t told = SysTick->VAL;    /* SysTick Current Value Register */
     do
@@ -51,8 +49,8 @@ void delay_us(uint16_t nus)
 
 void delay_ms(uint16_t nms)
 {
-    uint32_t tcnt = 0U;
-    uint32_t tnow = 0U;
+    uint32_t tcnt = 0;
+    uint32_t tnow = 0;
     uint32_t ticks = nms * fac_ms;
     uint32_t reload = SysTick->LOAD; /* SysTick Reload Value Register */
     uint32_t told = SysTick->VAL;    /* SysTick Current Value Register */
